requirements.txt
fastdtw
numpy
opencv-python
Pillow
pyqt5
scipy
